// Front-end/src/model/pet.ts
export interface Pet{
    name: string
    shelterName: string
    id: number
    species: string
    breed: string
    isBooked: boolean
    age: number
    behavior: string
    description: string
    houseTraining: boolean
    gender: string
    vaccination: string
    spaying: string
    neutering: string
}